+++ linkerd/tls/src/lib.rs
@@ -4,7 +4,6 @@ pub use linkerd_identity::LocalId;
 pub use rustls::TLSError as Error;
 
 pub mod client;
-mod conditional_accept;
 pub mod server;
 
 pub use self::{
